Added code to search the CollectiveMap and attempt to uniquely identify the detected objects. If the objects cannot be uniquely identified, they are added to the collective map as is. The added code has not been tested yet, but the program does compile and run identically to the last commit.

+++ collectiveIntelligence/src/QuadTreeGateway.scala
@@ -60,7 +60,7 @@ class QuadTreeGateway
 			outp
 	}
 
-	def range(radius: Int, x: Int, y: Int): List[Obstacle] =
+	def range(radius: Double, x: Double, y: Double): List[Obstacle] =
 	{
 		val input = qT.returnAll()
 		var output = Nil.asInstanceOf[List[Obstacle]]
